Moved Collision script & added to namespace
